[DEV-5856] Moving files around
