+++ usaspending_api/etl/management/commands/es_configure.py
@@ -4,7 +4,7 @@ import subprocess
 from django.core.management.base import BaseCommand
 
 from django.conf import settings
-from usaspending_api.etl.es_etl_helpers import VIEW_COLUMNS, AWARD_VIEW_COLUMNS
+from usaspending_api.etl.elasticsearch_loader_helpers.es_etl_helpers import VIEW_COLUMNS, AWARD_VIEW_COLUMNS
 
 CURL_STATEMENT = 'curl -XPUT "{url}" -H "Content-Type: application/json" -d \'{data}\''
